requirements.txt
python-telegram-bot
loguru
qrcode
Flask